+++ src/main/java/org/ton/ui/custom/layout/ResetBlockchainPaneController.java
@@ -3,12 +3,17 @@ package org.ton.ui.custom.layout;
 import javafx.fxml.FXML;
 import javafx.fxml.Initializable;
 import lombok.extern.slf4j.Slf4j;
+import org.apache.commons.io.FileUtils;
+import org.ton.actions.MyLocalTon;
+import org.ton.settings.MyLocalTonSettings;
 import org.ton.ui.custom.events.CustomEvent;
 import org.ton.ui.custom.events.event.CustomActionEvent;
 import org.ton.utils.Utils;
 
+import java.io.File;
 import java.net.URL;
 import java.util.ResourceBundle;
+import java.util.concurrent.ConcurrentLinkedQueue;
 
 import static com.sun.javafx.PlatformUtil.isWindows;
 import static org.ton.ui.custom.events.CustomEventBus.emit;
@@ -34,20 +39,41 @@ public class ResetBlockchainPaneController implements Initializable {
             Thread.sleep(100);
             if (isWindows()) {
                 if (Utils.doShutdown()) {
-                    log.info("restarting: cmd /c start java -jar {} restart", Utils.getMyPath());
-                    Runtime.getRuntime().exec("cmd /c start java -jar " + Utils.getMyPath() + " restart");
+                    Thread.sleep(200);
+                    MyLocalTon.getInstance().getSettings().setActiveNodes(new ConcurrentLinkedQueue<>());
+                    MyLocalTon.getInstance().saveSettingsToGson();
+                    deleteOldInstance();
+                    Thread.sleep(200);
                     System.exit(0);
                 }
             } else {
                 if (Utils.doShutdown()) {
-                    // works on linux
-                    log.info("restarting: java -jar {}", Utils.getMyPath());
-                    Runtime.getRuntime().exec("java -jar " + Utils.getMyPath() + " restart");
+                    Thread.sleep(200);
+                    MyLocalTon.getInstance().getSettings().setActiveNodes(new ConcurrentLinkedQueue<>());
+                    MyLocalTon.getInstance().saveSettingsToGson();
+                    deleteOldInstance();
+                    Thread.sleep(200);
                     System.exit(0);
                 }
             }
         } catch (Exception e) {
             log.error("Cannot restart MyLocalTon, error " + e.getMessage());
         }
+    }
+
+    private void deleteOldInstance() throws InterruptedException {
+        log.info("R E S E T T I N G");
+        FileUtils.deleteQuietly(new File(MyLocalTonSettings.MY_APP_DIR + File.separator + "MyLocalTonDB"));
+        FileUtils.deleteQuietly(new File(MyLocalTonSettings.MY_APP_DIR + File.separator + "genesis"));
+        FileUtils.deleteQuietly(new File(MyLocalTonSettings.MY_APP_DIR + File.separator + "node2"));
+        FileUtils.deleteQuietly(new File(MyLocalTonSettings.MY_APP_DIR + File.separator + "node3"));
+        FileUtils.deleteQuietly(new File(MyLocalTonSettings.MY_APP_DIR + File.separator + "node4"));
+        FileUtils.deleteQuietly(new File(MyLocalTonSettings.MY_APP_DIR + File.separator + "node5"));
+        FileUtils.deleteQuietly(new File(MyLocalTonSettings.MY_APP_DIR + File.separator + "node6"));
+        FileUtils.deleteQuietly(new File(MyLocalTonSettings.MY_APP_DIR + File.separator + "node7"));
+        FileUtils.deleteQuietly(new File(MyLocalTonSettings.MY_APP_DIR + File.separator + "templates"));
+        FileUtils.deleteQuietly(new File(MyLocalTonSettings.MY_APP_DIR + File.separator + "myLocalTon.log"));
+        Thread.sleep(500);
+        FileUtils.deleteQuietly(new File(MyLocalTonSettings.MY_APP_DIR + File.separator + "MyLocalTonDB"));
     }
 }